Decompose Rust gatekeeper into domain modules
